refactor kext invocation to separate cpp file
